Remove the Player's finish function.
They are more trouble than they are worth, it seems. The main problem is that
there is no way to time them sensibly -- if the first player's finish function
times out, we would have to change the scores completely. And then what to do if
the second player's finish function also times out?